DDT/requirements.txt
scikit_image==0.14.2
fire==0.1.3
numpy==1.16.2
torchvision==0.2.2.post3
torch==1.0.1.post2
Pillow==6.0.0
skimage==0.0
